requirements.txt
psycopg2==2.9.9
pandas==2.2.2
websocket-client==1.8.0
matplotlib==3.9.1